+++ packages/react-app/docs/CHANGELOG.md
@@ -1,11 +1,12 @@
 # CHANGELOG
 
-## 1.0.0-SNAPSHOT
+## 1.0.0-beta2-SNAPSHOT
 
 1. Use QA specific environment variables
 2. Add metadata to display preview in discord link
 3. Adjust hashID for discordjs webhook message
 4. Update ESLint config with cleaner controls and update build system
+5. Add bounty filters
 
 ## 1.0.0-beta (2021-08-12)